+++ tools/lkl/lib/utils.c
@@ -1,4 +1,3 @@
-#include <stdarg.h>
 #include <stdio.h>
 #include <string.h>
 #include <lkl_host.h>
@@ -157,54 +156,6 @@ void lkl_perror(char *msg, int err)
 	/* We need to use 'real' printf because lkl_host_ops.print can
 	 * be turned off when debugging is off. */
 	lkl_printf("%s: %s\n", msg, err_msg);
-}
-
-static int lkl_vprintf(const char *fmt, va_list args)
-{
-	int n;
-	char *buffer;
-	va_list copy;
-
-	if (!lkl_host_ops.print)
-		return 0;
-
-	va_copy(copy, args);
-	n = vsnprintf(NULL, 0, fmt, copy);
-	va_end(copy);
-
-	buffer = lkl_host_ops.mem_alloc(n + 1);
-	if (!buffer)
-		return -1;
-
-	vsnprintf(buffer, n + 1, fmt, args);
-
-	lkl_host_ops.print(buffer, n);
-	lkl_host_ops.mem_free(buffer);
-
-	return n;
-}
-
-int lkl_printf(const char *fmt, ...)
-{
-	int n;
-	va_list args;
-
-	va_start(args, fmt);
-	n = lkl_vprintf(fmt, args);
-	va_end(args);
-
-	return n;
-}
-
-void lkl_bug(const char *fmt, ...)
-{
-	va_list args;
-
-	va_start(args, fmt);
-	lkl_vprintf(fmt, args);
-	va_end(args);
-
-	lkl_host_ops.panic();
 }
 
 int lkl_sysctl(const char *path, const char *value)
